+++ libshvrpc/rpcerror.c
@@ -59,25 +59,27 @@ bool rpcerror_pack(cp_pack_t pack, rpcerrno_t errnum, const char *errmsg) {
 }
 
 static rpcerrno_t sanitize(rpcerrno_t errnum) {
-	return errnum > RPCERR_USER_CODE ? RPCERR_USER_CODE : errnum;
+	return errnum >= RPCERR_METHOD_CALL_EXCEPTION_BASE
+		? RPCERR_METHOD_CALL_EXCEPTION
+		: errnum > RPCERR_NOT_IMPLEMENTED ? RPCERR_UNKNOWN
+										  : errnum;
 }
 
 const char *rpcerror_str(rpcerrno_t errnum) {
 	static const char *const errname[] = {
 		[RPCERR_NO_ERROR] = "NoError",
-		[RPCERR_INVALID_REQUEST] = "InvalidRequest",
+		[1] = "InvalidRequest",
 		[RPCERR_METHOD_NOT_FOUND] = "MethodNotFound",
 		[RPCERR_INVALID_PARAM] = "InvalidParam",
-		[RPCERR_INTERNAL_ERR] = "InternalError",
-		[RPCERR_PARSE_ERR] = "ParseError",
-		[RPCERR_METHOD_CALL_TIMEOUT] = "MethodCallTimeout",
-		[RPCERR_METHOD_CALL_CANCELLED] = "MethodCallCancelled",
+		[4] = "InternalError",
+		[5] = "ParseError",
+		[RPCERR_USR1] = "ImplementationReserved1",
+		[RPCERR_USR2] = "ImplementationReserved2",
 		[RPCERR_METHOD_CALL_EXCEPTION] = "MethodCallException",
 		[RPCERR_UNKNOWN] = "Unknown",
 		[RPCERR_LOGIN_REQUIRED] = "LoginRequired",
 		[RPCERR_USER_ID_REQUIRED] = "UserIDRequired",
 		[RPCERR_NOT_IMPLEMENTED] = "NotImplemented",
-		[RPCERR_USER_CODE] = "UserCode",
 	};
 	return errname[sanitize(errnum)] ?: errname[RPCERR_UNKNOWN];
 }
@@ -85,23 +87,21 @@ const char *rpcerror_str(rpcerrno_t errnum) {
 const char *rpcerror_desc(rpcerrno_t errnum) {
 	static const char *const errdesc[] = {
 		[RPCERR_NO_ERROR] = "No error was encountered",
-		[RPCERR_INVALID_REQUEST] =
-			"The received value is not a valid RPC Request object",
+		[1] = "The received value is not a valid RPC Request object",
 		[RPCERR_METHOD_NOT_FOUND] =
 			"The requested method does not exist or is not available or "
 			"not accessible with given access level",
 		[RPCERR_INVALID_PARAM] = "Invalid request parameter",
-		[RPCERR_INTERNAL_ERR] = "Invernal RPC error",
-		[RPCERR_PARSE_ERR] = "Invalid ChainPack was received",
-		[RPCERR_METHOD_CALL_TIMEOUT] = "Method execution timed out",
-		[RPCERR_METHOD_CALL_CANCELLED] = "Method execution was canceled",
+		[4] = "Invernal RPC error",
+		[5] = "Invalid ChainPack was received",
+		[RPCERR_USR1] = "Implementation specific error 1",
+		[RPCERR_USR2] = "Implementation specific error 2",
 		[RPCERR_METHOD_CALL_EXCEPTION] = "Method execution exception",
 		[RPCERR_UNKNOWN] = "Unknown error",
 		[RPCERR_LOGIN_REQUIRED] =
 			"Method call attempted without previous successful login",
 		[RPCERR_USER_ID_REQUIRED] = "Method call requires UserID to be present",
 		[RPCERR_NOT_IMPLEMENTED] = "Method is valid but not implemented",
-		[RPCERR_USER_CODE] = "Used defined error",
 	};
 	return errdesc[sanitize(errnum)] ?: errdesc[RPCERR_UNKNOWN];
 }