+++ CHANGELOG.md
@@ -13,6 +13,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   and the number of records to be fetched, allowing better optimization
   from the application.
 
+### Removed
+- No longer valid RPC error codes (they were removed from standard).
+
 ### Fixed
 - RPC File's methods `sha1` and `crc` now correctly accept no parameter and
   single item in the list.